Fix bug in makeRandom and do some more general code clean up
The makeRandom() function was ignoring its argument and always returning 16
bytes of random data (this didn't matter because it was only ever called with
16 as its argument).

Take another look at this file and perform general clean up of comments and
go style.

Spotted another small bug where an error was set using

  err = errors.New()

but there was no return statement so the error was ignored.