static/index.html: Convert from single page app to tabbed app.
The diff for this is a bit ugly due to indentation changes, but
essentially each navbar tag that exists, was converted into a tab.  From
there, only one tab is shown at a time- this makes navigation generally
suck less, and is visually less painful.

In parallel, the markup for each tab was adjusted to be more consistent.